requirements.txt
argparse~=1.4.0
numpy~=2.0.1
pandas~=2.2.2
SQLAlchemy~=2.0.32
tqdm~=4.66.5
python-dateutil~=2.9.0.post0
matplotlib~=3.9.0
pyarrow
python-dotenv~=1.0.1
requests~=2.32.3